fix(build): split global type augmentation from module exports
