+++ src/bin/probot-run.ts
@@ -1,3 +1,3 @@
-import { run } from "../index.js";
+import { run } from "../exports.js";
 
 run(process.argv);
