+++ src/index.jsr.ts
@@ -1,0 +1,1 @@
+export * from "./exports.js";
