+++ src/types.ts
@@ -7,7 +7,7 @@ import type { LRUCache } from "lru-cache";
 import type { RedisOptions } from "ioredis";
 import type { Options as LoggingOptions } from "pino-http";
 
-import { Probot } from "./index.js";
+import { Probot } from "./exports.js";
 import { Context } from "./context.js";
 import { ProbotOctokit } from "./octokit/probot-octokit.js";
